created stakingSummary component, and exported claim function into the stakingHelper
